@@ -6,193 +6,202 @@ import (
 	"math/rand"
 )
 
-func DotProduct(x []float64, y []float64) float64 {
-	rv := 0.0
-	for i, v := range x {
-		rv += v * y[i]
+func dot_product(a []float64, b []float64) float64 {
+	n := len(a)
+	sum := 0.0
+	for i := range n {
+		sum += a[i] * b[i]
 	}
-	return rv
+	return sum
 }
 
-func Sigmoid(x float64) float64 {
-	return 1.0 / (1.0 + math.Exp(-x))
+func sigmoid(x float64) float64 {
+	return 1.0 / (1.0 + math.Exp(x))
 }
 
-func DerivativeSigmoid(x float64) float64 {
-	s := Sigmoid(x)
-	return s * (1.0 - s)
+func d_sigmoid(x float64) float64 {
+	sig := sigmoid(x)
+	return sig * (1.0 - sig)
 }
 
 type ActivationFunc func(float64) float64
+
 type Neuron struct {
-	weights          []float64
-	bias             float64
-	activation       ActivationFunc
-	derivative       ActivationFunc
-	activation_input float64
-	delta            float64
+	weights       []float64
+	wt_update     []float64
+	bias          float64
+	bias_update   float64
+	learning_rate float64
+	save_dot_prod float64
+	delta         float64
+
+	activation ActivationFunc
+	derivative ActivationFunc
 }
 
-func NewRandomNeuron(num_inputs int) *Neuron {
+func NewNeuron(num_inputs int) *Neuron {
 	n := Neuron{
-		weights:          make([]float64, num_inputs),
-		bias:             -1.0 + 2.0*rand.Float64(),
-		activation:       Sigmoid,
-		derivative:       DerivativeSigmoid,
-		activation_input: 0.0,
-		delta:            0.0,
+		weights:       make([]float64, num_inputs),
+		wt_update:     make([]float64, num_inputs),
+		bias:          -1.0 + 2.0*rand.Float64(),
+		bias_update:   0.0,
+		learning_rate: 0.5,
+		save_dot_prod: 0.0,
+		delta:         0.0,
+
+		activation: sigmoid,
+		derivative: d_sigmoid,
 	}
 
-	for i := 0; i < num_inputs; i++ {
+	for i := range num_inputs {
 		n.weights[i] = -1.0 + 2.0*rand.Float64()
+		n.wt_update[i] = 0.0
 	}
 
 	return &n
 }
 
-func NewNeuron(weights []float64, bias float64) *Neuron {
-	return &Neuron{
-		weights:          weights,
-		bias:             bias,
-		activation:       Sigmoid,
-		derivative:       DerivativeSigmoid,
-		activation_input: 0.0,
-		delta:            0.0,
-	}
+func (n *Neuron) Forward(inputs []float64) float64 {
+	n.save_dot_prod = dot_product(inputs, n.weights) + n.bias
+	return n.activation(n.save_dot_prod)
 }
 
-func (n *Neuron) Eval(inputs []float64) float64 {
-	n.activation_input = DotProduct(n.weights, inputs) + n.bias
-	return n.activation(n.activation_input)
+func (n *Neuron) BackProp(e float64, inputs []float64) {
+	n.delta = n.derivative(n.save_dot_prod) * e
+	for i, _ := range n.weights {
+		n.wt_update[i] += n.delta * n.learning_rate * inputs[i]
+	}
+	n.bias_update = n.delta * n.learning_rate
+}
+
+func (n *Neuron) UpdateWeights() float64 {
+	total := 0.0
+	for i, change := range n.wt_update {
+		n.weights[i] -= change
+		total += math.Abs(change)
+		n.wt_update[i] = 0.0
+	}
+	n.bias += n.bias_update
+	total += math.Abs(n.bias_update)
+	n.bias_update = 0.0
+
+	return total
 }
 
 type Layer struct {
-	neurons  []*Neuron
-	previous *Layer
-	input    []float64
+	neurons     []*Neuron
+	num_inputs  int
+	save_inputs []float64
 }
 
-func NewInputLayer(num_neurons int, num_inputs int) *Layer {
-	var layer Layer
-	for i := 0; i < num_neurons; i++ {
-		layer.neurons = append(layer.neurons, NewRandomNeuron(num_inputs))
+func NewLayer(num_neurons int, num_inputs int) *Layer {
+	x := Layer{
+		neurons:     make([]*Neuron, num_neurons),
+		num_inputs:  num_inputs,
+		save_inputs: make([]float64, num_inputs),
 	}
-	layer.previous = nil
-	layer.input = make([]float64, num_inputs)
 
-	return &layer
+	for i := range num_neurons {
+		x.neurons[i] = NewNeuron(num_inputs)
+	}
+
+	return &x
 }
 
-func NewLayer(num_neurons int, previous *Layer) *Layer {
-	num_inputs := len(previous.neurons)
-
-	var layer Layer
-	for i := 0; i < num_neurons; i++ {
-		layer.neurons = append(layer.neurons, NewRandomNeuron(num_inputs))
+func (x *Layer) Forward(input []float64) []float64 {
+	output := make([]float64, len(x.neurons))
+	for i, v := range input {
+		x.save_inputs[i] = v
 	}
-	layer.previous = previous
-	layer.input = make([]float64, num_inputs)
 
-	return &layer
+	for j, n := range x.neurons {
+		output[j] = n.Forward(input)
+	}
+
+	return output
 }
 
-func (layer *Layer) Forward(input []float64) []float64 {
-	if layer.previous == nil {
-		layer.input = input
-	} else {
-		layer.input = layer.previous.Forward(input)
-	}
+func (x *Layer) BackProp(e []float64) []float64 {
+	back := make([]float64, x.num_inputs)
 
-	num_neurons := len(layer.neurons)
-	rv := make([]float64, num_neurons)
-	for i := 0; i < num_neurons; i++ {
-		rv[i] = layer.neurons[i].Eval(layer.input)
-	}
-	return rv
-}
-
-// err = actual - expected
-func (layer *Layer) BackpropagateErrors(err []float64) {
-	eps := 0.0001
-
-	// First compute the error for each neuron of this layer
-	for i, n := range layer.neurons {
-		n.delta = n.derivative(n.activation_input) * err[i]
-	}
-
-	// back propagate the error
-	if layer.previous != nil {
-		num_layer_inputs := len(layer.neurons[0].weights)
-		propagation_err := make([]float64, num_layer_inputs)
-
-		for _, n := range layer.neurons {
-			total_wt := n.bias
-			for _, wt := range n.weights {
-				total_wt += wt
-			}
-
-			if total_wt > eps || total_wt < -eps {
-				for j, wt := range n.weights {
-					propagation_err[j] += n.delta * wt / total_wt
-				}
-			}
+	for i, n := range x.neurons {
+		n.BackProp(e[i], x.save_inputs)
+		for j, wt := range n.weights {
+			back[j] += wt * n.delta
 		}
-
-		layer.previous.BackpropagateErrors(propagation_err)
 	}
+	return back
+}
+
+func (x *Layer) UpdateWeights() float64 {
+	total := 0.0
+	for _, n := range x.neurons {
+		total += n.UpdateWeights()
+	}
+	return total
 }
 
 type Network struct {
-	num_inputs    int
-	layers        []*Layer
-	learning_rate float64
+	num_inputs int
+	layers     []*Layer
 }
 
-func NewNetwork(num_inputs int, learning_rate float64, layer_size []int) *Network {
-	var nn Network
-	nn.num_inputs = num_inputs
-	nn.learning_rate = learning_rate
-
-	{
-		layer := NewInputLayer(layer_size[0], num_inputs)
-		nn.layers = append(nn.layers, layer)
+func NewNetwork(num_inputs int, num_neurons ...int) *Network {
+	nn := Network{
+		num_inputs: num_inputs,
+		layers:     nil,
 	}
 
-	for i := 1; i < len(layer_size); i++ {
-		layer := NewLayer(layer_size[i], nn.layers[i-1])
-		nn.layers = append(nn.layers, layer)
+	prev := num_inputs
+
+	for _, a := range num_neurons {
+		x := NewLayer(a, prev)
+		nn.layers = append(nn.layers, x)
+		prev = a
 	}
+
 	return &nn
 }
 
 func (nn *Network) Forward(input []float64) []float64 {
+	vec := make([]float64, nn.num_inputs)
+	copy(vec, input)
 	num_layers := len(nn.layers)
-	return nn.layers[num_layers-1].Forward(input)
+	for i := 0; i < num_layers; i++ {
+		vec = nn.layers[i].Forward(vec)
+	}
+	return vec
 }
 
-func (nn *Network) BackpropagateErrors(err []float64) {
+func (nn *Network) BackProp(e []float64) {
+	vec := make([]float64, nn.num_inputs)
+	copy(vec, e)
 	num_layers := len(nn.layers)
-	nn.layers[num_layers-1].BackpropagateErrors(err)
+	for i := num_layers - 1; i >= 0; i-- {
+		vec = nn.layers[i].BackProp(vec)
+	}
 }
 
 func (nn *Network) UpdateWeights() float64 {
-	rv := 0.0
-	for _, layer := range nn.layers {
-		for _, n := range layer.neurons {
-			for i, _ := range n.weights {
-				change := nn.learning_rate * layer.input[i] * n.delta
-				rv += change * change
-				n.weights[i] -= change
-			}
-
-			{
-				change := nn.learning_rate * n.delta
-				rv += change * change
-				n.bias -= change
-			}
-		}
+	total := 0.0
+	for _, x := range nn.layers {
+		total += x.UpdateWeights()
 	}
-	return rv
+	return total
+}
+
+func (nn *Network) TrainOneData(input []float64, expect []float64) float64 {
+	actual := nn.Forward(input)
+	num_outputs := len(expect)
+	e := make([]float64, num_outputs)
+	total_error := 0.0
+	for i := range num_outputs {
+		e[i] = expect[i] - actual[i]
+		total_error += math.Abs(e[i])
+	}
+
+	nn.BackProp(e)
+	return total_error
 }
 
 type NNData struct {
@@ -201,23 +210,12 @@ type NNData struct {
 }
 
 func (nn *Network) Train(data []NNData, num_epochs int) {
-	for epoch := 0; epoch < num_epochs; epoch++ {
+	for epoch := range num_epochs {
 		total_error := 0.0
-		total_change := 0.0
 		for _, d := range data {
-			out := nn.Forward(d.Input)
-			fmt.Printf("%6.2f %6.2f %6.2f : %6.2f %6.2f %6.2f\n",
-				out[0], out[1], out[2], d.Output[0], d.Output[1], d.Output[2])
-			err := make([]float64, nn.num_inputs)
-			for i, v := range out {
-				err[i] = v - d.Output[i]
-				total_error += err[i] * err[i]
-			}
-			nn.BackpropagateErrors(err)
-			total_change += nn.UpdateWeights()
+			total_error += nn.TrainOneData(d.Input, d.Output)
 		}
-		fmt.Println(epoch, total_error, total_change)
-
+		change := nn.UpdateWeights()
+		fmt.Printf("%5d %20.10f %20.10f\n", epoch, total_error, change)
 	}
-
 }
